fix(OBAKitCore): move UIKit-dependent files to OBAKit with conditional compilation
- Move 9 files from OBAKitCore to OBAKit:
  ActivityIndicatedButton, ListKitExtensions, AutoLayoutExtensions,
  ArrivalDepartureDrivenUI, DepartureTimeBadge, PaddingLabel,
  ProminentButton, ImageBadgeRenderer, UIViewPreview
- Add import OBAKitCore to moved files that reference OBAKitCore types
- Split UIKitExtensions.swift: iOS-only extensions moved to
  OBAKit/Extensions/UIKitExtensions+OBAKitCore.swift
- Add #if os(iOS) guards around Autolayoutable protocol conformance
  and layout direction extensions in UIKitExtensions.swift
- Add #if os(watchOS) alternatives for contrastingTextColor
