+++ OBAKit/UI/ProminentButton.swift
@@ -8,6 +8,7 @@
 //
 
 import UIKit
+import OBAKitCore
 
 /// A button with a background color, designed to give the button more of a tappable affordance.
 public class ProminentButton: UIButton {
